requirements.txt
albumentations==2.0.8
joblib==1.5.1
matplotlib==3.7.5
monai==1.5.0
nibabel==5.3.2
numpy==1.26.4
opencv-python==4.6.0.66
pandas==2.3.1
pillow==11.3.0
protobuf==5.29.5
python-dotenv==1.0.1
scikit-learn==1.3.2
scipy==1.13.1
seaborn==0.13.2
synapseclient==4.9.0
timm==1.0.19
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.4
#transformers==4.41.0
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==1.26.20
